+++ portal/models/forms/provider_cache.go
@@ -1,0 +1,10 @@
+// Copyright (c) 2015-2022 CloudJ Technology Co., Ltd.
+
+package forms
+
+type ClearProviderCacheForm struct {
+	BaseForm
+
+	Source  string `json:"source" form:"source" binding:"required"`
+	Version string `json:"version" form:"version" binding:"required"`
+}