+++ runner/api/v1/handler/provider_cache.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2015-2022 CloudJ Technology Co., Ltd.
+
+package handler
+
+import (
+	"cloudiac/configs"
+	"cloudiac/runner"
+	"cloudiac/runner/api/ctx"
+	"net/http"
+	"path/filepath"
+	"strings"
+)
+
+func RunClearProviderCache(c *ctx.Context) {
+	req := runner.RunClearProviderCacheReq{}
+	if err := c.BindJSON(&req); err != nil {
+		c.Error(err, http.StatusBadRequest)
+		return
+	}
+
+	count := strings.Count(req.Source, "/")
+	providerCachePath := configs.Get().Runner.AbsProviderCachePath()
+	if count == 2 {
+		if ok, _ := runner.DeleteProviderCache(providerCachePath, req.Source, req.Version); ok { //nolint
+			return
+		}
+	} else if count == 1 {
+		if ok, _ := runner.DeleteProviderCache(filepath.Join(providerCachePath, "registry.terraform.io"), req.Source, req.Version); ok { //nolint
+			return
+		}
+	}
+}