+++ scripts/updatedocs/main.go
@@ -16,8 +16,8 @@ import (
 )
 
 const (
-	version    = "v1.1.0"
-	webVersion = "v1.1.0"
+	version    = "v1.2.2"
+	webVersion = "v1.2.2"
 )
 
 type FileReplacer struct {